// src/interfaces/Permissions.ts
export enum PermissionList {
  VIEW_PRODUCT = 'VIEW_PRODUCT',
  VIEW_CATEGORY = 'VIEW_CATEGORY',
  LIST_PRODUCTS = 'LIST_PRODUCTS',
  LIST_CATEGORIES = 'LIST_CATEGORIES',
  CREATE_PRODUCT = 'CREATE_PRODUCT',
  CREATE_CATEGORY = 'CREATE_CATEGORY',
  UPDATE_PRODUCT = 'UPDATE_PRODUCT',
  UPDATE_CATEGORY = 'UPDATE_CATEGORY',
  DELETE_PRODUCT = 'DELETE_PRODUCT',
  DELETE_CATEGORY = 'DELETE_CATEGORY',
  CREATE_ORDER = 'CREATE_ORDER',
  VIEW_ORDER = 'VIEW_ORDER',
  LIST_ORDERS = 'LIST_ORDERS',
  UPDATE_ORDER_STATUS = 'UPDATE_ORDER_STATUS',
  CREATE_USER = 'CREATE_USER',
  UPDATE_USER = 'UPDATE_USER',
  DELETE_USER = 'DELETE_USER',
  LIST_USERS = 'LIST_USERS',
  CREATE_CART = 'CREATE_CART',
  UPDATE_CART = 'UPDATE_CART',
  DELETE_CART = 'DELETE_CART',
  VIEW_CART = 'VIEW_CART',
}
export interface RolePermissions {
  admin: PermissionList[];
  manager: PermissionList[];
  user: PermissionList[];
}
export type PermissionRoleType = 'admin' | 'manager' | 'user';